refactor(utils): consolidate utility imports into a single index file
This commit reorganizes the utility functions by creating an `index.ts` in the `utils` directory. This change simplifies import statements across the codebase, allowing for cleaner and more concise imports of utility functions like `createPrevWeekDate` and `createStringFromDate`. The adjustments were made in various components, enhancing maintainability and readability of the code without altering any functionality.
