+++ src/features/meal-form/components/MealForm.tsx
@@ -8,7 +8,7 @@ import { MealCategoryName } from '@/API'
 import { useCurrentDateStore } from '../../../stores/currentDate'
 import { useMealRecordsStore } from '../../../stores/mealRecords'
 import { useNutritionNumbersStore } from '../../../stores/nutritionNumbers'
-import { createStringFromDate } from '../../../utils/createStringFromDate'
+import { createStringFromDate } from '../../../utils'
 import { fetchMealRecords } from '../api'
 import { FormsType } from '../types'
 import { createInitialFormValues, createSumNutritionValues } from '../utils'
