+++ src/features/meal-form/components/MealFormAccordionItem.tsx
@@ -5,7 +5,7 @@ import { useEffect } from 'react'
 import { DialogSaved } from '../../../components/DialogSaved'
 import { useCurrentDateStore } from '../../../stores/currentDate'
 import { useMealRecordsStore } from '../../../stores/mealRecords'
-import { createStringFromDate } from '../../../utils/createStringFromDate'
+import { createStringFromDate } from '../../../utils'
 import { FormsType } from '../types'
 import { createFoodInitialValues, saveMealRecord } from '../utils'
 import { MealFormFields } from './MealFormFields'
